Move IronMQ driver to new namespace
